Move version info from Makefile to git hooks
This commit removes a script from the Makefile that generates
the commit date and version number and stores them in the file
VERSION.tex (which then is included in the main document). This
information is now generated by the git hooks:
- post-commit
- post-merge
- post-checkout

This approach is motivated by Brent Longborough's gitinfo package.