Clean up and tweak project Makefile
Some targets were being unnecessarily rebuilt; this commit makes a
few changes to the Makefile that should prevent that. In particular,
paper.pdf depended on the appendix's plots and that's been fixed.

After this commit we still sometimes leave the pdf files older than
their dependencies if latexmk doesn't decide to run, but I'm reluctant
to force-update the timestamp for the files. It seems like that's a
case where we should manually run pdflatex just to make sure everything
has been processed.

Fixup last commit