[#49] Import latex-tools repository as subdirectory
We're going to use the version_git bash script and makefile snippet
to automatically generate a version and commit stamp for the pdfs
produced in this repository. To make life easier for potential
researchers not familiar with Git, I'm just going to add the current
version of the latex-tools repository as a subdirectory instead of
using a submodule. This commit adds that subdirectory.